backend/requirements.txt
Flask
requests
numpy
opencv-python
Pillow
pytesseract
